+++ lib/lucid/cli/options.rb
@@ -28,16 +28,16 @@ module Lucid
       largest = LUCID_FORMATS.keys.map{|s| s.length}.max
       FORMAT_LIST = (LUCID_FORMATS.keys.sort.map do |key|
         "  #{key}#{' ' * (largest - key.length)} : #{LUCID_FORMATS[key][1]}"
-      end) + ["Use --format rerun --out rerun.txt to write out failing",
-              "specs. You can rerun them with lucid @rerun.txt.",
-              "FORMAT can also be the fully qualified class name of",
+      end) + ['Use --format rerun --out rerun.txt to write out failing',
+              'specs. You can rerun them with lucid @rerun.txt.',
+              'FORMAT can also be the fully qualified class name of',
               "your own custom formatter. If the class isn't loaded,",
-              "Lucid will attempt to require a file with a relative",
-              "file name that is the underscore name of the class name.",
-              "  Example: --format Formatter::WordDoc",
-              "With that, Lucid will look for formatter/word_doc.rb",
-              "You can place the file with this relative path",
-              "underneath your common/support directory or anywhere",
+              'Lucid will attempt to require a file with a relative',
+              'file name that is the underscore name of the class name.',
+              '  Example: --format Formatter::WordDoc',
+              'With that, Lucid will look for formatter/word_doc.rb',
+              'You can place the file with this relative path',
+              'underneath your common/support directory or anywhere',
               "on Ruby's LOAD_PATH."
       ]
       PROFILE_SHORT_FLAG = '-p'
@@ -83,23 +83,24 @@ module Lucid
         @args.extend(::OptionParser::Arguable)
 
         @args.options do |opts|
-          opts.banner = ["Lucid: Test Description Language Execution Engine",
-                         "Usage: lucid [options] [ [FILE|DIR|URL][:LINE[:LINE]*] ]+", "", ""
+          opts.banner = ['Lucid: Test Description Language Execution Engine',
+                         'Usage: lucid [options] [ [FILE|DIR|URL][:LINE[:LINE]*] ]+', '', ''
           ].join("\n")
 
-          opts.on("--library-path PATH", "Location of spec project library files.") do |path|
+          opts.on('--library-path PATH', 'Location of spec project library files.') do |path|
             @options[:library_path] = path
           end
 
-          opts.on("--spec-type TYPE", "The file type (extension) for Lucid specifications.") do |type|
-            @options[:spec_type] = type
+          opts.on('--spec-type TYPE', 'The file type (extension) for Lucid specifications.') do |type|
+            #@options[:spec_type] = type
+            @options[:spec_type] << type
           end
 
-          opts.on("--driver-file FILE", "The file for Lucid to connect to an execution library.") do |file|
+          opts.on('--driver-file FILE', 'The file for Lucid to connect to an execution library.') do |file|
             @options[:driver_file] = file
           end
 
-          opts.separator ""
+          opts.separator ''
 
           opts.on('-r LIBRARY|DIR', '--require LIBRARY|DIR',
                   'Require files before executing the features. If this option',
@@ -107,8 +108,9 @@ module Lucid
                   'the features will be loaded automatically. Automatic loading',
                   'is disabled when this option is specified. That means all',
                   'loading becomes explicit.',
-                  "Files under directories named \"support\" will always be",
-                  'loaded first.',
+                  'Assuming a default specs repo configuration, files under',
+                  "directories named \"common\\support\" will always be loaded",
+                  'before any others.',
                   'This option can be specified multiple times.') do |v|
             @options[:require] << v
             if(Lucid::JRUBY && File.directory?(v))
@@ -147,10 +149,10 @@ module Lucid
           end
 
           opts.on('-n NAME', '--name NAME',
-                  'Lucid will only execute features or abilities that match with the name",
-                  "provided. The match can be done on partial information. If this option",
-                  "is provided multiple times, then the match will be performed against",
-                  "each set of provided names.'
+                  'Lucid will only execute features or abilities that match with the name',
+                  'provided. The match can be done on partial information. If this option',
+                  'is provided multiple times, then the match will be performed against',
+                  'each set of provided names.'
           ) do |v|
             @options[:name_regexps] << /#{v}/
           end
@@ -203,73 +205,73 @@ module Lucid
 
           opts.separator ''
 
-          opts.on("-c", "--[no-]color",
-                  "Specifies whether or not to use ANSI color in the output. If this",
-                  "option is not specified, Lucid makes the decision on colored output",
-                  "based on your platform and the output destination."
+          opts.on('-c', '--[no-]color',
+                  'Specifies whether or not to use ANSI color in the output. If this',
+                  'option is not specified, Lucid makes the decision on colored output',
+                  'based on your platform and the output destination.'
           ) do |v|
             Lucid::Term::ANSIColor.coloring = v
           end
 
-          opts.on("-m", "--no-multiline",
-                  "Lucid will not print multiline strings and tables under steps.") do
+          opts.on('-m', '--no-multiline',
+                  'Lucid will not print multiline strings and tables under steps.') do
             @options[:no_multiline] = true
           end
 
-          opts.on("-s", "--no-source",
-                  "Lucid will not print the file and line of the test definition with the steps.") do
+          opts.on('-s', '--no-source',
+                  'Lucid will not print the file and line of the test definition with the steps.') do
             @options[:source] = false
           end
 
-          opts.on("-i", "--no-matchers",
-                  "Lucid will not print matchers for pending steps.") do
+          opts.on('-i', '--no-matchers',
+                  'Lucid will not print matchers for pending steps.') do
             @options[:matchers] = false
           end
 
-          opts.on("-I", "--matchers-type TYPE",
-                  "Use different matcher type (Default: regexp).",
-                  "Available types:",
+          opts.on('-I', '--matchers-type TYPE',
+                  'Use different matcher type (Default: regexp).',
+                  'Available types:',
                   *Lucid::InterfaceRb::RbLanguage.cli_matcher_type_options
           ) do |v|
             @options[:matcher_type] = v.to_sym
           end
 
-          opts.on("-q", "--quiet", "Alias for --no-matchers --no-source.") do
+          opts.on('-q', '--quiet', 'Alias for --no-matchers --no-source.') do
             @quiet = true
           end
 
-          opts.on("-S", "--strict", "Fail if there are any undefined or pending steps.") do
+          opts.on('-S', '--strict', 'Fail if there are any undefined or pending steps.') do
             @options[:strict] = true
           end
 
-          opts.on("-w", "--wip", "Fail if there are any passing scenarios.") do
+          opts.on('-w', '--wip', 'Fail if there are any passing scenarios.') do
             @options[:wip] = true
           end
 
-          opts.on("-g", "--guess", "Guess best match for ambiguous steps.") do
+          opts.on('-g', '--guess', 'Guess best match for ambiguous steps.') do
             @options[:guess] = true
           end
 
-          opts.on("-x", "--expand", "Expand Scenario Outline tables in output.") do
+          opts.on('-x', '--expand', 'Expand Scenario Outline tables in output.') do
             @options[:expand] = true
           end
 
-          opts.separator ""
+          opts.separator ''
 
-          opts.on("--testdefs DIR", "Lucid will Write test definition metadata to the DIR.") do |dir|
+          opts.on('--testdefs DIR', 'Lucid will write test definition metadata to the DIR.') do |dir|
             @options[:testdefs] = dir
           end
 
           if(Lucid::JRUBY)
-            opts.on("-j DIR", "--jars DIR",
-                    "Load all the jars under the specified directory.") do |jars|
+            opts.on('-j DIR', '--jars DIR',
+                    'Load all the jars under the specified directory.') do |jars|
               Dir["#{jars}/**/*.jar"].each {|jar| require jar}
             end
           end
 
-          opts.on("--i18n LANG",
-                  "List keywords for a particular language.",
-                  %{Run with "--i18n help" to see all languages}) do |lang|
+          opts.on('--i18n LANG',
+                  'List keywords for a particular language.',
+                  'Run with "--i18n help" to see all languages') do |lang|
             if lang == 'help'
               list_languages_and_exit
             else
@@ -310,14 +312,14 @@ module Lucid
         rescue OptionParser::InvalidOption
           if $!.to_s =~ /invalid option\:\s+((?:-)?-\S+)/
             puts "You specified an invalid option: #{$1}"
-            puts "Please run lucid --help to see the list of available options."
+            puts 'Please run lucid --help to see the list of available options.'
           end
 
         rescue OptionParser::MissingArgument
           if $!.to_s =~ /missing argument\:\s+((?:-)?-\S+)/
             puts "You specified an valid option (#{$1}), but with an invalid argument."
-            puts "Make sure you are providing the expected argument for the option."
-            puts "Run lucid --help to see the list of available options."
+            puts 'Make sure you are providing the expected argument for the option.'
+            puts 'Run lucid --help to see the list of available options.'
           end
 
           Kernel.exit(1)
@@ -417,6 +419,7 @@ module Lucid
       def reverse_merge(other_options)
         @options = other_options.options.merge(@options)
         @options[:require] += other_options[:require]
+
         @options[:excludes] += other_options[:excludes]
         @options[:name_regexps] += other_options[:name_regexps]
         @options[:tag_expressions] += other_options[:tag_expressions]
@@ -471,9 +474,9 @@ module Lucid
           :name_regexps     => [],
           :env_vars         => {},
           :diff_enabled     => true,
-          :spec_type        => "",
-          :library_path     => "",
-          :driver_file      => ""
+          :spec_type        => %w(feature spec story),
+          :library_path     => '',
+          :driver_file      => ''
         }
       end
     end