+++ lib/lucid/runtime/facade.rb
@@ -4,7 +4,7 @@ module Lucid
   class Runtime
     # This is what a programming language will consider to be a runtime.
     #
-    # It's a thin class that directs the handul of methods needed by the
+    # It's a thin class that directs the specific methods needed by the
     # programming languages to the right place.
     class Facade
       extend Forwardable
@@ -25,21 +25,20 @@ module Lucid
         :invoke,
         :load_code_language
 
-      # Returns a Lucid::AST::Table for +text_or_table+, which can either
-      # be a String:
+      # Returns a Lucid::AST::Table which can either be a String:
       #
       #   table(%{
-      #     | account | description | amount |
-      #     | INT-100 | Taxi        | 114    |
-      #     | CUC-101 | Peeler      | 22     |
+      #     | study   | phase  |
+      #     | Test-01 | I      |
+      #     | Test-02 | II     |
       #   })
       #
       # or a 2D Array:
       #
       #   table([
-      #     %w{ account description amount },
-      #     %w{ INT-100 Taxi        114    },
-      #     %w{ CUC-101 Peeler      22     }
+      #     %w{ study phase },
+      #     %w{ Test-01 I },
+      #     %w{ Test-02 II }
       #   ])
       #
       def table(text_or_table, file=nil, line_offset=0)
@@ -50,8 +49,7 @@ module Lucid
         end
       end
 
-      # Returns AST::DocString for +string_without_triple_quotes+.
-      #
+      # Returns AST::DocString
       def doc_string(string_without_triple_quotes, content_type='', line_offset=0)
         AST::DocString.new(string_without_triple_quotes,content_type)
       end