+++ spec/spec_helper.rb
@@ -3,11 +3,17 @@ $LOAD_PATH.unshift(File.dirname(__FILE__))
 require 'coveralls'
 Coveralls.wear!
 
-if ENV['SIMPLECOV']
-  require 'simplecov'
-  SimpleCov.start do
-    command_name 'rspec'
-  end
+SimpleCov.add_filter '/spec'
+
+SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter[
+  SimpleCov::Formatter::HTMLFormatter,
+  Coveralls::SimpleCov::Formatter
+]
+
+SimpleCov.at_exit do
+  SimpleCov.result.format!
+  SimpleCov.minimum_coverage 60
+  SimpleCov.maximum_coverage_drop 5
 end
 
 require 'lucid'